Add breakpoint_site and stoppoint_collection
Breakpoint site will be responsible for tracking breakpoints at a
location level, so at an adress level. Stoppoint is meant to act as a
template to any object that might stop our debuggee's execution:
breakpoint sites, source-level breakpoints, and watchpoints. It is in
this class that we will set up basic breakpoint interaction like adding,
removing, enabling and iterating through a collection of breakpoints